+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- `v.0.0.5` Badge component
+  - Require at least one of `label` or `icon`
+
 ## 0.10.2 - 2024-06-06
 ### Fixed
 - `v.0.0.4` Chip component